add e2e test for inmemory crd channel
