+++ test/common/creation.go
@@ -17,7 +17,7 @@ limitations under the License.
 package common
 
 import (
-	"github.com/knative/eventing/test/base"
+	"github.com/knative/eventing/test/base/resources"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -31,21 +31,27 @@ var rbacAPIGroup = rbacv1.SchemeGroupVersion.Group
 var rbacAPIVersion = rbacv1.SchemeGroupVersion.Version
 
 // CreateChannelOrFail will create a Channel Resource in Eventing.
-// TODO(Fredy-Z): This is a workaround when there are both provisioner and Channel CRDs in this repo.
-//                isCRD needs to be deleted when the provisioner implementation is removed.
 func (client *Client) CreateChannelOrFail(name string, channelTypeMeta *metav1.TypeMeta, provisionerName string) {
 	namespace := client.Namespace
 	switch channelTypeMeta.Kind {
-	case base.ChannelKind:
-		channel := base.Channel(name, provisionerName)
+	case resources.ChannelKind:
+		channel := resources.Channel(name, provisionerName)
 		channels := client.Eventing.EventingV1alpha1().Channels(namespace)
 		channel, err := channels.Create(channel)
 		if err != nil {
 			client.T.Fatalf("Failed to create channel %q: %v", name, err)
 		}
 		client.Cleaner.AddObj(channel)
-	case base.KafkaChannelKind:
-		channel := base.KafkaChannel(name)
+	case resources.InMemoryChannelKind:
+		channel := resources.InMemoryChannel(name)
+		channels := client.Eventing.MessagingV1alpha1().InMemoryChannels(namespace)
+		channel, err := channels.Create(channel)
+		if err != nil {
+			client.T.Fatalf("Failed to create %q %q: %v", channelTypeMeta.Kind, name, err)
+		}
+		client.Cleaner.AddObj(channel)
+	case resources.KafkaChannelKind:
+		channel := resources.KafkaChannel(name)
 		channels := client.Kafka.MessagingV1alpha1().KafkaChannels(namespace)
 		channel, err := channels.Create(channel)
 		if err != nil {
@@ -66,10 +72,10 @@ func (client *Client) CreateChannelsOrFail(names []string, channelTypeMeta *meta
 func (client *Client) CreateSubscriptionOrFail(
 	name, channelName string,
 	channelTypeMeta *metav1.TypeMeta,
-	options ...base.SubscriptionOption,
+	options ...resources.SubscriptionOption,
 ) {
 	namespace := client.Namespace
-	subscription := base.Subscription(name, channelName, channelTypeMeta, options...)
+	subscription := resources.Subscription(name, channelName, channelTypeMeta, options...)
 
 	subscriptions := client.Eventing.EventingV1alpha1().Subscriptions(namespace)
 	// update subscription with the new reference
@@ -85,7 +91,7 @@ func (client *Client) CreateSubscriptionsOrFail(
 	names []string,
 	channelName string,
 	channelTypeMeta *metav1.TypeMeta,
-	options ...base.SubscriptionOption,
+	options ...resources.SubscriptionOption,
 ) {
 	for _, name := range names {
 		client.CreateSubscriptionOrFail(name, channelName, channelTypeMeta, options...)
@@ -95,7 +101,7 @@ func (client *Client) CreateSubscriptionsOrFail(
 // CreateBrokerOrFail will create a Broker or fail the test if there is an error.
 func (client *Client) CreateBrokerOrFail(name, provisionerName string) {
 	namespace := client.Namespace
-	broker := base.Broker(name, provisionerName)
+	broker := resources.Broker(name, provisionerName)
 
 	brokers := client.Eventing.EventingV1alpha1().Brokers(namespace)
 	// update broker with the new reference
@@ -114,9 +120,9 @@ func (client *Client) CreateBrokersOrFail(names []string, provisionerName string
 }
 
 // CreateTriggerOrFail will create a Trigger or fail the test if there is an error.
-func (client *Client) CreateTriggerOrFail(name string, options ...base.TriggerOption) {
+func (client *Client) CreateTriggerOrFail(name string, options ...resources.TriggerOption) {
 	namespace := client.Namespace
-	trigger := base.Trigger(name, options...)
+	trigger := resources.Trigger(name, options...)
 
 	triggers := client.Eventing.EventingV1alpha1().Triggers(namespace)
 	// update trigger with the new reference
@@ -132,10 +138,10 @@ func (client *Client) CreateCronJobSourceOrFail(
 	name,
 	schedule,
 	data string,
-	options ...base.CronJobSourceOption,
+	options ...resources.CronJobSourceOption,
 ) {
 	namespace := client.Namespace
-	cronJobSource := base.CronJobSource(name, schedule, data, options...)
+	cronJobSource := resources.CronJobSource(name, schedule, data, options...)
 
 	cronJobSources := client.Eventing.SourcesV1alpha1().CronJobSources(namespace)
 	// update cronJobSource with the new reference
@@ -146,13 +152,13 @@ func (client *Client) CreateCronJobSourceOrFail(
 	client.Cleaner.AddObj(cronJobSource)
 }
 
-// CreateContainerSourceOrFail will create a ContainerSource.
+// CreateContainerSourceOrFail will create a ContainerSource or fail the test if there is an error.
 func (client *Client) CreateContainerSourceOrFail(
 	name string,
-	options ...base.ContainerSourceOption,
+	options ...resources.ContainerSourceOption,
 ) {
 	namespace := client.Namespace
-	containerSource := base.ContainerSource(name, options...)
+	containerSource := resources.ContainerSource(name, options...)
 
 	containerSources := client.Eventing.SourcesV1alpha1().ContainerSources(namespace)
 	// update containerSource with the new reference
@@ -167,7 +173,7 @@ func (client *Client) CreateContainerSourceOrFail(
 func WithService(name string) func(*corev1.Pod, *Client) error {
 	return func(pod *corev1.Pod, client *Client) error {
 		namespace := pod.Namespace
-		svc := base.Service(name, pod.Labels)
+		svc := resources.Service(name, pod.Labels)
 
 		svcs := client.Kube.Kube.CoreV1().Services(namespace)
 		if _, err := svcs.Create(svc); err != nil {
@@ -199,14 +205,14 @@ func (client *Client) CreatePodOrFail(pod *corev1.Pod, options ...func(*corev1.P
 // cluster-admin role.
 func (client *Client) CreateServiceAccountAndBindingOrFail(saName, crName string) {
 	namespace := client.Namespace
-	sa := base.ServiceAccount(saName, namespace)
+	sa := resources.ServiceAccount(saName, namespace)
 	sas := client.Kube.Kube.CoreV1().ServiceAccounts(namespace)
 	if _, err := sas.Create(sa); err != nil {
 		client.T.Fatalf("Failed to create service account %q: %v", saName, err)
 	}
 	client.Cleaner.Add(coreAPIGroup, coreAPIVersion, "serviceaccounts", namespace, saName)
 
-	crb := base.ClusterRoleBinding(saName, crName, namespace)
+	crb := resources.ClusterRoleBinding(saName, crName, namespace)
 	crbs := client.Kube.Kube.RbacV1().ClusterRoleBindings()
 	if _, err := crbs.Create(crb); err != nil {
 		client.T.Fatalf("Failed to create cluster role binding %q: %v", crName, err)
