+++ test/test_images/logevents/kodata/VENDOR-LICENSE
